Update patch set 4
Patch Set 4:

Also, in general, I believe AERs are supposed to be reset by reading the corresponding Get Log Page, so we should really not be removing events from the queue when the AER command is received but actually when the log page is read.  In other words, if a NS change notice is triggered and the user submits more AER commands without reading the log page, we should keep completing the AER with the same event.

Patch-set: 4
Label: Code-Review=0