Update patch set 7
Patch Set 7:

I changed not only the type (int -> bool) but also the name, for example, chap_required -> require_chap based on the chat with Daniel.

before this patch, noun + verb + ed
after this patch, verb + object.

I would like to adopt the better one and restoring the previous names are OK for me.

In subsequent patches, I just want to fix inconsistency of names based on the result of this patch.

Thank you for your review.

Patch-set: 7