rte_vhost/virtio: increase max memory regions limit to 128
Increase the limit from 8 to 128.

The DPDK memory allocator creates a separate hugetlbfs
file for each mapped hugepage. This is fine for most cases,
but not for virtio initiator. For virtio, we have to send
a descriptor for each of these files to the vhost process, so it
can remap the memory and perform DMA on it. When using 2MB hugepages,
this implies sharing possibly hundreds of descriptors per
connection, and managing possibly thousands of them per vhost.
There's an artificial MAX_MEMORY_REGIONS(8) limit can be always
increased, but we are still likely to hit the max open descriptors
limit for a single process. There's also a hardcoded limit on
a number of descriptors sent via sendmsg() - currently 253 in Linux.

Due to this DPDK behavior, virtio initiator was practically
limited to 1GB hugepages. With this patch it could be
theoretically used with 128 2MB hugepages, 256MB memory.
And it will work for some properly prepared storage setups.
But that's not the use case behind this patch. The next
patch in this series enables DPDK to combine physically
contiguous pages (hugetlbfs files) into a single, bigger
hugetlbfs file. On heavily-fragmented machines, however, there
might be no physically contiguous pages available, and DPDK
may still create as many files as pages. But that's just extemely
unlikely. The proper use case for this patch is to add virtio
initiator support for small-memory-footprint applications
like containers, requiring much less than 1GB of memory and hence
unable to use 1GB hugepages. Depending on the machine setup,
*some* of these initiators might use more than 8 memory regions.

We increase the max memory region limit to 128 just to be safe.

Side note:
sizeof(VhostUserMsg) is now ~4,5KB; it's always allocated on
the stack.

Change-Id: I68d774fa7c128a49427cc725c236e8ba968a8894
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>