blob: allocate persist_ctx before CLEAN check
Future patches will need to check if an existing
persist is in progress *before* checking for CLEAN,
for supporting multiple persist operations
outstanding at once for a single blob.

So prep for the patch a bit here, by allocating the
persist_ctx first, then freeing it if the state is
CLEAN.  This enables the future patch since the
persist_ctx is the thing we will be queueing if
necessary.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Ic06a087990d7140c5c0f6011b51b16e7d8e1695b