Update patch set 5
Patch Set 5: Code-Review+1

I agree with both of you.

It does seem incorrect that we can send nopin even before login.  We should fix this.  But as Ziye mentioned - the existing code before this patch has the same behavior.

So I think this patch is OK.  Ziye - could you create a separate patch later that ensures we do not send nopin for a connection that has not logged in yet?  I think you can check conn->full_feature - if it is zero, then we should not do any nop related work.

I will mark this patch +1 for now.  Shuhei, if you agree with my proposal, you may change your vote to +2 and then I will revise my vote to +2 as well.

Thanks,

-Jim

Patch-set: 5