bdev: add a current_ticks member for bdev stats
This patch is to add a current_ticks member for the
bdev spdk_bdev_io_stat structure. In this way, it
can keep a track on the current ticks when other
callers to get the stats when necessary. By dosing
so, we can remove the operation to zero out the
captured stats while using the current ticks to
calculate the I/O statistics like IOPS.

Change-Id: I6f25a287eef0cbd8bd9b19ad3b2e69b4e6c54475
Signed-off-by: GangCao <gang.cao@intel.com>