Update patch set 4
Patch Set 4: Code-Review-1

(4 comments)

Hi Lance,

I'd really like to keep this patch just focused in removing the exit() calls in the libraries.  Let's defer the positive/negative errno differentiation.  I'm still not convinced this differentiation will be helpful and would like to separate that discussion from the one that we all agree with which is removing the exit() calls from the libraries.

Thanks,

-Jim

Patch-set: 4
Label: Code-Review=-1