iscsi: Separate initializing and configuring portal group at boot
Initialization of portal and portal group must be done before
starting iSCSI target application, but Configuration of portal and
portal group can be done after starting iSCSI target application by
JSON-RPC.

This patch separates initialization of portal and portal group and
configuration of portal and portal group at SPDK booting.

This is a preparation for JSON config file. The purpose of this patch
series is to separate global parameters and config parameters for
iSCSI.

Change-Id: I8b01c5d44b24b31221d8b1d9a844b25dd5e060cb
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>