Update patch set 10
Patch Set 10:

> There's a few more mentions of unittest.sh that need to be fixed
 > up:
 > 
 > - README.md
 > - doc/getting_started.md
 > - .travis.yml
 > 
 > It might make sense to keep a stub unittest.sh that just locates
 > rootdir and calls $rootdir/test/unit/unittest.sh (I would
 > appreciate this too, since typing ./unittest.sh is muscle memory
 > for me now...), but I'm open to arguments against this.

Thanks for the pointers to additional references.
The only argument I have against the stub unittest.sh is that it seems to go against the strict hierarchy structure defined by the tiger team. Plus, it would be easy to create that file locally for your convenience and leave it on your machine without checking it in.

Patch-set: 10
Label: Code-Review=0