iscsi: Separate initializing and configuring portal group at boot
To utilize current code as much as possible to support JSON config
file, separate the following:
- initialize iSCSI subsystem by using global parameters in the config
  file
- configure portal group, initiator group, and target node by using
  config information in the config file

This patch take a task to complete this goal.

Initializing the portal list and the portal group list must be done
even if the config file does not have any information, but
configuring portals and portal groups will bo NOP if the config
file does not have any information.

Change-Id: I8b01c5d44b24b31221d8b1d9a844b25dd5e060cb
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>