iscsi: Separate initializing and configuring initiator group at boot
To utilize current code as much as possible to support JSON config
file, separate the following:
- initialize iSCSI subsystem by using global parameters in the config
  file
- configure portal group, initiator group, and target node by using
  config information in the config file

This patch take a task to complete this goal.

Initializing the initiator group list must be done even if the config
file does not have any information, but configuring initiator groups
will bo NOP if the config file does not have any information.

Change-Id: I434f0abf2d8a9b015e847f05a58a45139761d11a
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>