iscsi: Separate initializing and configuring target node at boot
To utilize current code as much as possible to support JSON config
file, separate the following:
- initialize iSCSI subsystem by using global parameters in the config
  file
- configure portal group, initiator group, and target node by using
  config information in the config file

This patch takes a task to complete this goal.

Initializing the target node list must be done even if the config
file does not have any information, but configuring target nodes
will bo NOP if the config file does not have any information.

Change-Id: I7467b26526a316d684a9add7fa40a11627356892
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>