Update patch set 2
Patch Set 2:

My first goal is to use this UUID for NVMe-oF namespace identification.

However, I agree that it sounds like a good idea to add it as a bdev alias (or use uuid as name and the friendly name as an alias, if that makes more sense).

Patch-set: 2
Reviewer: Daniel Verkamp <1010525@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0