[RFC] bdev: add crypto vbdev
Patch Set 2 adds support for separate src/dst on encryption but requires
a few changes to DPDK that aren't in this patch.  Will review those
separately.  Patch Set 1 will work with stock DPDK with the config
changes listed below (and the addtl lib) however will encrypt 'in place'
in the provided source buffer.

Not ready for review but posting to support Q&A/discussion.  Todo items
include:

* For it to build: (need to intgrate/automate)
- Copy and build this library: https://github.com/01org/intel-ipsec-mb/archive/v0.47.zip
- export AESNI_MULTI_BUFFER_LIB_PATH=/home/peluse/intel-ipsec-mb-0.47
- cp ~/intel-ipsec-mb-0.47/libIPSec_MB.a /home/peluse/spdk/dpdk/build/lib/libIPSec_MB.a
- edit dpdk/config/common_spdk
   CONFIG_RTE_LIBRTE_PMD_AESNI_MB=y
   CONFIG_RTE_LIBRTE_CRYPTODEV=y
   CONFIG_RTE_LIBRTE_REORDER=y

* Complete inline TODO items
* Discuss various #define values
* Discuss QAT vs this virtual PMD
* Identifiy optimizations
* Remove a bunch of debug log messages
* Discuss "cypto in place" limitation of the AESNI PMD
* Decide on what to do with assertions in bdevio (data compare)
* Discuss other possible issues w/in palce encryption (vhost, etc)
* ?

Change-Id: If518c3df8e74e00efa18afdf194824c5e69778fc
Signed-off-by: Paul Luse <paul.e.luse@intel.com>