bdev: add crypto vbdev
WIP: refactoring one function, do not review. Updated & rebased to test execution
in CI. There's also one change to teset_env.c that doesn't belong in this patch,
will remove after CI testing.

QAT can be included in the code (but not makefile) and marked as experimental
until we are ready to test in CI.  It works well on 2 systems but is a big PITA to get
the HW setup and configured for use with DPDK (IMHO).

Have tested (both aesni and QAT seprately and concurrrently) on underlying NVMe devices
with bdevio and a bdevperf script that runs IOs from 512B to 128K each with Q depths from
1 to 512 in powers of 2 for 30 seconds each run.

Change-Id: If518c3df8e74e00efa18afdf194824c5e69778fc
Signed-off-by: Paul Luse <paul.e.luse@intel.com>