Update patch set 1
Patch Set 1:

For NVMf target crash, I checked the code, it seems that
        thread = 0x0
        ch = 0x0
in _spdk_put_io_channel.
Which means: (1) We are successfully sends the message to that thread or core; (2) but when we execute that, the thread info is not valid.

So there is a question, why different threads get the io channel of other thread? In NVMe-oF target, the io channel of name space is getting from spdk_get_bdev_io_channel.

Patch-set: 1
Reviewer: Ziye Yang <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0