Update patch set 2
Patch Set 2: Code-Review+2

Changing vote to +2.  This patch is functionally OK - but would like to see a follow-up patch to address my two comments.

This allows us to commit another 7 patches after this one that were OK and avoid an extra rebase of all of those patches.

Patch-set: 2
Label: Code-Review=+2