Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

Hi Tomasz, 

Logic looks correct to me but advancing the current pointer may be easier to understand. Please check the following.

I'm not sure the name curr_payload is right and if mine is wrong or yours is better, it's OK for me.

Thanks,
Shuhei

Patch-set: 3
Label: Code-Review=-1