Update patch set 63
Patch Set 63: Code-Review-1

(5 comments)

This is looking really good!

One big thing that is missing is recovery if the system fails in the middle of the snapshotting operation.  We can add that in a later patch.  For now, please just add in the commit message that recovery is not covered in this patch and that it will be added later.

Patch-set: 63
Label: Code-Review=-1