Update patch set 23
Patch Set 23:

Thank you Shuhei.

I know what your mean. Adding this function is called by RPC that is executed just by one core, hence I think that is safe for multiple threads and remove the mutex. Since that is also a public function, we need considering the safe for multiple threads.

So what do you think which is the much better solution? Adding a lock between RPC output context or as your means, using a asynchronous function.  For this target, it's better to get this information with low latency since 'iostat' tool need to get the I/O information as soon as possible.

I look forward to meeting you in Beijing this week and thanks a lot.

Patch-set: 23