test: add more info to the test log
Not sure how much this will help as there was already compiler
info there, but added uname, moved it up in the log a little
but (I think) and put some markers before and afer the info.

This is to help folks determine what OS & tool chain versions
the CI tests systems are using.

Change-Id: I49d2de09a8d8e5fbcdfa07ad2ce20adfc250d4e3
Signed-off-by: Paul Luse <paul.e.luse@intel.com>
Reviewed-on: https://review.gerrithub.io/404416
Tested-by: SPDK Automated Test System <sys_sgsw@intel.com>
Reviewed-by: Daniel Verkamp <daniel.verkamp@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>