Update patch set 1
Patch Set 1:

> These are intentionally not in unittest.sh, since they need DPDK
 > set up to work correctly (all of the other unit tests do not depend
 > on the env layer).

That makes sense. In the interest of conforming to the new script hierarchy, would it be ok if we moved the directory to test/unit, but did not call the tests from unittest.sh to preserve that barrier?

Patch-set: 1
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0