Update patch set 11
Patch Set 11:

Tomasz, thanks! I met with the AESNI maintainers today and for this PMD at least they explain that enqueue doesn't actually start any real work, it just puts the ops in a buffer.  When you dequeue is when the crypto operation actually happens.  I suspect this is different for real HW and the reason they kept two calls was to keep the generic API in tact.

Anyway, see the code I have in there now.  The only reason I'd want to keep dequeueing until done is if I do it inline (the option I just added). Otherwise, if using a poller I believe I want to enqueue everything on submission (which may include a dequeue operation if the PMD is full, that code is in there now, but in the poller I don't think I want to sit in there and run dequeue operations until the thing is empty. I suppose I could - either would work I guess its a matter of testing to see which performs better.  As enqueue and dequeue are on the same thread maybe it does make more sense to dequeue everything but in general I guess I'm so used to this kind of routine being like a high priority INT (which its not of course) that I'm just inclined to do what work is available and then get out, come back later and do what's available then.  Still lots of experimentation to do, thanks for all the input!

Patch-set: 11
Label: Code-Review=0