event/subsystem: Check completion of related subsystem_inits before each subsystem_init
When JSON config file is supported, each subsystem initialization will be done by
JSON-RPC. Hence when a subsystem is initialized, it must be confirmed that
all subsystems on which the subsystem depends are initialized already.

This patch adds necessary functions and they are inserted into the current code
to validate before JSON-CONFIG file is supported.

Change-Id: I70c20e17f7f2535b3cd9ea48c0eec9ffc54c0ed4
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>