Update patch set 5
Patch Set 5: Code-Review-1

Until this incorporates VPP, this patch would be best suited to be based on master - rather than VPP series. It would be preferable to have this test up and running beforehand.

Another patch could be created that adds support for VPP, and indeed it could be placed on top of VPP series.

Patch-set: 5
Label: Code-Review=-1