log: add a print_level field to log components
This will allow tweaking individual log component levels beyond just
enabled/disabled for debug and info logs.

Currently, this isn't exposed in the API; doing so will require an API
break in both the C interface and the RPC.  For now, just map the
trace_flag APIs so that enabling a flag causes it to print DEBUG and
higher and disabling a flag causes it to print INFO and higher.

Change-Id: I6144b91e92c668338c86b9c6e76421e740fea89b
Signed-off-by: Daniel Verkamp <daniel.verkamp@intel.com>