Update patch set 2
Patch Set 2:

The build_pool implementation of the pool currently captures all of the output of the post-processing step (autorun_post.py) so that the script checked into the SPDK repo doesn't need to handle it (and so output is captured even if the script fails part of the way through).  However, I think this is a reasonable change to make so that we have just the test execution information in a separate file.

Patch-set: 2