Update patch set 51
Patch Set 51:

Talked about this with Tomek.  A lot of the functionality for this is in blobstore, not lvol/lvol_bdev so adding a bunch of unit tests here may not add a lot of value.

I still think it would be good to add code specifically for new lvol.c code for creating snapshots and clones where it has to check the name (just like a new lvol).  So some unit tests for that would be good.

Patch-set: 51