Update patch set 3
Patch Set 3:

As long as I discussed shortly with Pawel, my subsequent patches will be similar to Pawel's. Please take this as an alternative idea. I will wait for Pawel's patches.

For this patch, I feel differentiation between post-subsys-init-RPCs and pre-subsys-init-RPCs is useful as I discussed with Jim and Daniel.

Patch-set: 3
Label: Code-Review=0