rpc: Introduce state mask to filter RPCs only for subsystem initialization
This is is to differentiate between RPCs that is allowed before subsystem
initialization and RPCs that is allowed after subsystem initialization.

Former are existing JSON RPCs and later are new for subsystem initialization.

The following is a short description of the design.

Add state_mask to struct spdk_rpc_method and add g_rpc_state to the RPC server.

Define two bits for the state_mask:
- RPC_STATE_PRE_SUBSYSTEM_INIT  0x1
- RPC_STATE_POST_SUBSYSTEM_INIT 0x2

g_rpc_state is set to either one by spdk_rpc_set_state(state_mask).

If the command line option '-j' is specified, g_rpc_state is initialized to
RPC_STATE_PRE_SUBSYSTEM_INIT. Otherwise, g_rpc_state is initialized to
RPC_STATE_POST_SUBSYSTEM_INIT.

In spdk_jsonrpc_handler(), each method's state_mask and g_rpc_state are
ANDed and the result is not 0, the method is allowed. Otherwise, the method is
not allowed.

To change g_rpc_state to RPC_STATE_POST_SUBSYSTEM_INIT, a new RPC
enable_post_si_rpc is introduced. This is allowed only when g_rpc_state is
RPC_STATE_PRE_SUBSYSTEM_INIT.

Change-Id: I4fa9c22b64e2fdbc9b9fdb2c47f0018f73f84f7e
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>